IEEE 802.15.4 does _not_ use network byte order - i.e. it transmits "backwards" (least significant byte and least significant bit first).
The packet_sink class was therefore shifting bits into the incorrect
position. After 32 samples are received, chip 0 would end up being in
bit position 31. It should be in bit position 0.

Furthermore, in spite of the reference to "CMOS Architectures.." the
chip table has been replaced with one that conforms to the IEEE
standard (versions 2006, 2009, and 2011).

What surprises me is that this code actually worked for the RIME stack,
although there were no unit tests to verify aside from running the
simulation.

Prior to this fix, the packet_sink block would not work intercepting
ZigBee traffic, claiming that the first SFD "byte" (actually nibble) was
incorrect.

A unit test is included for validation.